Add Piston + keypad input
Add a dependency on piston_window [1] to create a window
and an event loop. 0xFX0A (Await Keypress) is implemented
with presses coming from the keyboard.

May want to refactor how "awaiting" is implemented since
the current solution of returing a bool from Cpu::execute()
does not feel super clean.

[1] https://github.com/PistonDevelopers/piston_window