Add piston_window event handling
Adds support for key presses (a key map holds the current state of
keys) -- due to this map removed the Keymap enum and switched to a
simple index into the map.

Adds support for dragging and dropping ROMs onto the window to start
playing that ROM.

NOTE: Key presses are still not handled by the emulator.