Add drawing support for HTML5
Added support for drawing to an HTML5 Canvas; a method was added to
chip8.rs that allows us to check if a pixel is on (true) or not (false).
This saves us from having to copy a 64*32 pixel array between Rust and
JS (since we can not pass a reference across the boundary).

Currently the game view is redrawn on every tick which is fairly
inefficient and causes visible slow downs in movement. This will be
resolved in a follow on CL.